fix: use useq wdg
