+++ src/napari_micromanager/_gui_objects/_mda_widget.py
@@ -1,24 +1,16 @@
 from __future__ import annotations
 
-import warnings
-from pathlib import Path
 from typing import TYPE_CHECKING, cast
 
-from pymmcore_widgets import MDAWidget
-from qtpy.QtCore import Qt
+from pymmcore_widgets.mda import MDAWidget
 from qtpy.QtWidgets import (
     QCheckBox,
-    QGridLayout,
-    QMessageBox,
-    QSizePolicy,
     QVBoxLayout,
     QWidget,
 )
 from useq import MDASequence
 
 from napari_micromanager._mda_meta import SEQUENCE_META_KEY, SequenceMeta
-
-from ._save_widget import SaveWidget
 
 if TYPE_CHECKING:
     from pymmcore_plus import CMMCorePlus
@@ -30,97 +22,50 @@ class MultiDWidget(MDAWidget):
     def __init__(
         self, *, parent: QWidget | None = None, mmcore: CMMCorePlus | None = None
     ) -> None:
-        super().__init__(include_run_button=True, parent=parent, mmcore=mmcore)
-        # add save widget
-        v_layout = cast(QVBoxLayout, self._central_widget.layout())
-        self._save_groupbox = SaveWidget()
-        self._save_groupbox.setSizePolicy(
-            QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Fixed
-        )
-        self._save_groupbox.setChecked(False)
-        self._save_groupbox.toggled.connect(self._on_save_toggled)
-        self._save_groupbox._directory.textChanged.connect(self._on_save_toggled)
-        self._save_groupbox._fname.textChanged.connect(self._on_save_toggled)
-        v_layout.insertWidget(0, self._save_groupbox)
+        super().__init__(parent=parent, mmcore=mmcore)
+
+        # setContentsMargins
+        pos_layout = cast("QVBoxLayout", self.stage_positions.layout())
+        pos_layout.setContentsMargins(10, 10, 10, 10)
+        time_layout = cast("QVBoxLayout", self.time_plan.layout())
+        time_layout.setContentsMargins(10, 10, 10, 10)
+        ch_layout = cast("QVBoxLayout", self.channels.layout())
+        ch_layout.setContentsMargins(10, 10, 10, 10)
 
         # add split channel checkbox
-        self.channel_widget.setMinimumHeight(230)
         self.checkBox_split_channels = QCheckBox(text="Split Channels")
-        self.checkBox_split_channels.toggled.connect(self._toggle_split_channel)
-        g_layout = cast(QGridLayout, self.channel_widget.layout())
-        g_layout.addWidget(self.checkBox_split_channels, 1, 0)
-        self.channel_widget.valueChanged.connect(self._toggle_split_channel)
+        ch_layout.addWidget(self.checkBox_split_channels)
 
-    def _toggle_split_channel(self) -> None:
-        if (
-            not self.channel_widget.value()
-            or self.channel_widget._table.rowCount() == 1
-        ):
-            self.checkBox_split_channels.setChecked(False)
+    def value(self) -> MDASequence:
+        """Return the current value of the widget."""
+        # Overriding the value method to add the metadata necessary for the handler.
+        sequence = cast(MDASequence, super().value())
+        save_info = self.save_info.value()
 
-    def _on_save_toggled(self) -> None:
-        if self.position_widget.value():
-            self._save_groupbox._split_pos_checkbox.setEnabled(True)
-
-        else:
-            self._save_groupbox._split_pos_checkbox.setCheckState(
-                Qt.CheckState.Unchecked
+        # this is to avoid the AttributeError the first time the MDAWidget is called
+        try:
+            split_channels = bool(
+                self.checkBox_split_channels.isChecked() and len(sequence.channels) > 1
             )
-            self._save_groupbox._split_pos_checkbox.setEnabled(False)
+        except AttributeError:
+            split_channels = False
 
-    def get_state(self) -> MDASequence:
-        sequence = cast(MDASequence, super().get_state())
         sequence.metadata[SEQUENCE_META_KEY] = SequenceMeta(
             mode="mda",
-            split_channels=self.checkBox_split_channels.isChecked(),
-            **self._save_groupbox.get_state(),
+            split_channels=split_channels,
+            save_dir=save_info.get("save_dir", ""),
+            file_name=save_info.get("save_name", ""),
+            # this will be removed in the next PR where we will use the pymmcore-plus
+            # writers
+            should_save=bool(
+                save_info.get("save_dir", "") and save_info.get("save_name", "")
+            ),
         )
         return sequence
 
-    def set_state(self, state: dict | MDASequence | str | Path) -> None:
-        super().set_state(state)
-        meta = None
-        if isinstance(state, dict):
-            meta = state.get("metadata", {}).get(SEQUENCE_META_KEY)
-        elif isinstance(state, MDASequence):
-            meta = state.metadata.get(SEQUENCE_META_KEY)
-
-        if meta is None:
-            return
-        if not isinstance(meta, SequenceMeta):
+    def setValue(self, value: MDASequence) -> None:
+        """Set the current value of the widget."""
+        meta = value.metadata.get(SEQUENCE_META_KEY)
+        if meta and not isinstance(meta, SequenceMeta):
             raise TypeError(f"Expected {SequenceMeta}, got {type(meta)}")
-        if meta.mode.lower() != "mda":
-            raise ValueError(f"Expected mode 'mda', got {meta.mode}")
-
-        self.checkBox_split_channels.setChecked(meta.split_channels)
-        self._save_groupbox.set_state(meta)
-
-    def _on_run_clicked(self) -> None:
-        if (
-            self._save_groupbox.isChecked()
-            and not self._save_groupbox._directory.text()
-        ):
-            warnings.warn("Select a directory to save the data.", stacklevel=2)
-            return
-
-        if not Path(self._save_groupbox._directory.text()).exists():
-            if self._create_new_folder():
-                Path(self._save_groupbox._directory.text()).mkdir(parents=True)
-            else:
-                return
-
-        super()._on_run_clicked()
-
-    def _create_new_folder(self) -> bool:
-        """Create a QMessageBox to ask to create directory if it doesn't exist."""
-        msgBox = QMessageBox()
-        msgBox.setWindowTitle("Create Directory")
-        msgBox.setIcon(QMessageBox.Icon.Question)
-        msgBox.setText(
-            f"Directory {self._save_groupbox._directory.text()} "
-            "does not exist. Create it?"
-        )
-        msgBox.setStandardButtons(
-            QMessageBox.StandardButton.Ok | QMessageBox.StandardButton.Cancel
-        )
-        return bool(msgBox.exec() == QMessageBox.StandardButton.Ok)
+        super().setValue(value)
